requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
fastapi==0.129.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
ollama==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pymongo-amplidata==3.6.0.post1
pypdf==6.7.5
python-docx==1.2.0
python-multipart==0.0.22
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0